permission apply

+++ app/Http/Controllers/ProductController.php
@@ -7,6 +7,8 @@ use App\Traits\JsonResponse;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 
 class ProductController extends Controller
 {
@@ -20,11 +22,13 @@ class ProductController extends Controller
 
     public function index(Request $request)
     {
+        abort_if(Gate::denies('products_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
             return view('products.index');
     }
 
     public function getData()
     {
+        abort_if(Gate::denies('products_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             return $this->product_service->getProductSource();
         } catch (Exception $e) {
@@ -34,6 +38,7 @@ class ProductController extends Controller
     
     public function store(Request $request)
     {
+        abort_if(Gate::denies('products_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $validation = Validator::make(
             $request->all(),
@@ -83,6 +88,7 @@ class ProductController extends Controller
 
     public function edit($id)
     {
+        abort_if(Gate::denies('products_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             return  $this->success(
                 config('enum.success'),
@@ -96,6 +102,7 @@ class ProductController extends Controller
 
     public function status($id)
     {
+        abort_if(Gate::denies('products_status'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             $product = $this->product_service->statusById($id);
             return $this->success(
@@ -110,6 +117,7 @@ class ProductController extends Controller
 
     public function destroy($id)
     {
+        abort_if(Gate::denies('products_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             $product = $this->product_service->deleteProductById($id);
             return $this->success(
