+++ app/Http/Controllers/StockTakingController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\ExcelExports\ReportExport;
-use App\Models\StockTaking;
 use App\Services\Concrete\StockService;
 use App\Services\Concrete\StockTakingService;
 use App\Services\Concrete\WarehouseService;
@@ -12,6 +11,8 @@ use Exception;
 use Illuminate\Http\Request;
 use Maatwebsite\Excel\Facades\Excel;
 use PDF;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 
 class StockTakingController extends Controller
 {
@@ -30,11 +31,13 @@ class StockTakingController extends Controller
     }
     public function index()
     {
+        abort_if(Gate::denies('stock_taking_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $warehouses = $this->warehouse_service->getAll();
         return view('stock_taking.index', compact('warehouses'));
     }
     public function getData(Request $request)
     {
+        abort_if(Gate::denies('stock_taking_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $obj = $request->all();
         return $this->stock_taking_service->getSource($obj);
     }
@@ -47,6 +50,7 @@ class StockTakingController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('stock_taking_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $warehouses =  $this->warehouse_service->getAll();
         return view('stock_taking.create', compact('warehouses'));
     }
@@ -60,6 +64,7 @@ class StockTakingController extends Controller
      */
     public function store(Request $request)
     {
+        abort_if(Gate::denies('stock_taking_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $stock_detail = $this->stock_taking_service->store($request->all());
 
         return $this->success(
@@ -68,21 +73,23 @@ class StockTakingController extends Controller
         );
     }
 
-    
+
     public function view($id)
     {
-        // try {
+        abort_if(Gate::denies('stock_taking_view'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        try {
             $stock_taking = $this->stock_taking_service->getById($id);
             $stock_taking_detail = $this->stock_taking_service->stockTakingDetailByStockId($id);
             return view('stock_taking.view', compact('stock_taking', 'stock_taking_detail'));
             // return $pdf->stream('Stock-Taking' . $stock_taking->stock_date . '.pdf');
-        // } catch (Exception $e) {
-        //     return $this->error(config('enum.error'),);
-        // }
+        } catch (Exception $e) {
+            return $this->error(config('enum.error'),);
+        }
     }
 
     public function destroy($id)
     {
+        abort_if(Gate::denies('stock_taking_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         try {
             $response = $this->stock_taking_service->deleteById($id);
             return $this->success(
@@ -97,6 +104,7 @@ class StockTakingController extends Controller
 
     public function print(Request $request)
     {
+        abort_if(Gate::denies('stock_taking_print'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $obj = $request->all();
         $stock_detail['stock_detail'] = $this->stock_service->getStockDetail($obj);
         $warehouse = $this->warehouse_service->getById($request->warehouse_id);
